There are several fixes that were required for getting json working.
The wizard UI didn't send the data type. It was always sending 0. Moved the logic to the endpoint to look for a "/" in the root element, if its there it counts it as json.

Fixed the the detail viewer, the related map is not done yet for json, I made a check to see if the map was there before doing the work for in in api-details

There were issues in doing the queries to find a json document to view in order to send to marklogic to review. Such as the uploader. it was trying to get the namespace of json documents. It doesn't do that any more.

The get by root element was also not working for Json documents. I created a fix for that as well.

Also the modules were not load so I created a task that loads them in.